fix: better init flow and default sets
